third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/credentialless/shared-worker.tentative.https-expected.txt

This is a testharness.js-based test.
PASS shared-worker
PASS fetch same-origin
PASS fetch same-origin + credentialless worker
PASS fetch cross-origin
FAIL fetch cross-origin + credentialless worker assert_equals: coep:none =>  expected (undefined) undefined but got (string) "cross_origin"
Harness: the test ran to completion.
